+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ Next version
   with ``fetch=True``. Users should only register the concrete plugin directly
   for fetching, otherwise duplicated plugins are very likely while rendering.
 - Added ``usedforsecurity=False`` to our usage of ``hashlib.md5``.
+- Bumped the minimum django-tree-queries version to 0.21 and used the
+  ``TreeAdmin`` and ``AncestorFilter`` from there.
 
 
 5.4 (2025-06-19)